% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipotleFunctions.R
\name{orderBowl}
\alias{orderBowl}
\title{Build Bowl and Total Cost}
\usage{
orderBowl(meat, rice, bean, top, extraTop)
}
\arguments{
\item{meat}{A numeric value for meat order}

\item{rice}{A numeric value for rice order}

\item{bean}{A numeric value for bean order}

\item{top}{A numeric value for toppings order}

\item{extraTop}{A numeric value for extra toppings order}
}
\value{
Total Cost of Bowl
}
\description{
Takes in numeric values for the respective menu items and prints receipt
}
